Create Certifications component and add to App layout
